changed exploration constant to double

+++ cpp_modules/src/mcts/MctsMeta.h
@@ -4,5 +4,5 @@
 
 #ifndef MCTSMETA_H
 #define MCTSMETA_H
-const float EXPLORATION = sqrt(2);
+const double EXPLORATION = sqrt(2);
 #endif //MCTSMETA_H
